requirements.txt
python-telegram-bot==20.7
SQLAlchemy==2.0.23
python-dotenv==1.0.1
aiohttp==3.9.5
pytz==2023.3.post1
matplotlib==3.8.2
fastapi==0.110.0
uvicorn==0.24.0.post1
